Refactor listings View component
