+++ components/listings/Listing/Properties/styles.js
