+++ components/listings/Listing/Thumbnail/index.js
