+++ components/listings/Listing/Thumbnail/styles.js
