+++ containers/listings/Listing/index.js
@@ -1,0 +1,8 @@
+import View from '@/components/listings/Listing'
+import Loader from './Loader'
+
+export default function ListingApp({id, ...props}) {
+  return (
+    <Loader id={id}>{({data}) => data && <View {...data} {...props} />}</Loader>
+  )
+}
